Sap and Session improved with methods to find and verify SapComponent's.
Methods that allow to find within a GuiModalWindow that's within a
GuiSession for example. Just like FindById or FindByText but limiting
the scope within a session's child component.

Code organization: Centralizing to generic static methods with the bulk
of code within Sap class and using those methods to build facilitators
within Session.