Bug with Worksheet.UsedRange fixed
Just using the UsedRange property from Worksheet object of the Excel
Object Model has shown to be unreliable. Some formats can lead to an
incorrect result of used cells in the sheet. As an example I had a
sheet with hundreds of lines but the UsedRange property retrieved just
the first one.

As a working solution was added some new logic to try to calculate
using the Range.Find method. Also was added a new property
(WorksheetFullRange) capable of retrieving all the range that excel
actualy loads in memory and makes available to the user type in (all
the sheet cells). Some methods like the private auxiliary
Excel.FixReadAllInput() were changed to use the Full Range instead of
the calculated Used Range. This way it will pass the full loaded range
to the GetUsedRange method that will calculate the used range in a
cell-by-cell manner using the continuity as criteria to go to the next
row or column when parsing.